fix: put image on newline for narrower screens
When viewing the readme on a narrow viewport, the text in the "How does it work" section gets messed up by the height of the image due to the image being on the same line as the text. Now the image will be on its own line, allowing the text to wrap to a new line without getting all wonky.